Refactor import paths in logger, test adapters, app config, and metrics modules to remove 'src' prefix for consistency.
